// src/modules/user/user.type.ts
export type UserType = {
  id: string;
  email: string;
  first_name: string;
  last_name: string;
  password: string;
  website: string;
  profession: string;
  address_line_1: string;
  address_line_2: string;
  city: string;
  zip_code: number;
  country: string;
  state: string;
  verified: boolean;
};